created feedback_validator.py & feedback_validator_test.py by refactoring prod_validator.py
